requirements.txt
pylabrobot>=0.5.0
unitelabs-tecanfluentcontrol>=1.0.0
sila2>=0.10.0
asyncio>=3.4.3
aiohttp>=3.8.0
typing-extensions>=4.0.0
pathlib>=1.0.1